refactor(utils): move inputValidators to utils
